+++ src/network.cxx
@@ -92,7 +92,7 @@ float get_current_rate(network *net)
     size_t batch_num = get_current_batch(net);
     int i;
     float rate;
-    if (batch_num < net->burn_in) return net->learning_rate * pow((float)batch_num / net->burn_in, net->power);
+    if ((int)batch_num < net->burn_in) return net->learning_rate * pow((float)batch_num / net->burn_in, net->power);
     switch (net->policy) {
         case CONSTANT:
             return net->learning_rate;
@@ -101,7 +101,7 @@ float get_current_rate(network *net)
         case STEPS:
             rate = net->learning_rate;
             for(i = 0; i < net->num_steps; ++i){
-                if(net->steps[i] > batch_num) return rate;
+                if(net->steps[i] > (int)batch_num) return rate;
                 rate *= net->scales[i];
             }
             return rate;
